+++ subsys/bluetooth/shell/bredr.c
@@ -45,74 +45,6 @@ NET_BUF_POOL_DEFINE(data_pool, 1, DATA_BREDR_MTU, BT_BUF_USER_DATA_MIN,
 NET_BUF_POOL_DEFINE(sdp_client_pool, CONFIG_BT_MAX_CONN,
 		    SDP_CLIENT_USER_BUF_LEN, BT_BUF_USER_DATA_MIN, NULL);
 
-#if defined(CONFIG_BT_RFCOMM)
-
-static struct bt_sdp_attribute spp_attrs[] = {
-	BT_SDP_NEW_SERVICE,
-	BT_SDP_LIST(
-		BT_SDP_ATTR_SVCLASS_ID_LIST,
-		BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 3),
-		BT_SDP_DATA_ELEM_LIST(
-		{
-			BT_SDP_TYPE_SIZE(BT_SDP_UUID16),
-			BT_SDP_ARRAY_16(BT_SDP_SERIAL_PORT_SVCLASS)
-		},
-		)
-	),
-	BT_SDP_LIST(
-		BT_SDP_ATTR_PROTO_DESC_LIST,
-		BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 12),
-		BT_SDP_DATA_ELEM_LIST(
-		{
-			BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 3),
-			BT_SDP_DATA_ELEM_LIST(
-			{
-				BT_SDP_TYPE_SIZE(BT_SDP_UUID16),
-				BT_SDP_ARRAY_16(BT_SDP_PROTO_L2CAP)
-			},
-			)
-		},
-		{
-			BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 5),
-			BT_SDP_DATA_ELEM_LIST(
-			{
-				BT_SDP_TYPE_SIZE(BT_SDP_UUID16),
-				BT_SDP_ARRAY_16(BT_SDP_PROTO_RFCOMM)
-			},
-			{
-				BT_SDP_TYPE_SIZE(BT_SDP_UINT8),
-				BT_SDP_ARRAY_8(BT_RFCOMM_CHAN_SPP)
-			},
-			)
-		},
-		)
-	),
-	BT_SDP_LIST(
-		BT_SDP_ATTR_PROFILE_DESC_LIST,
-		BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 8),
-		BT_SDP_DATA_ELEM_LIST(
-		{
-			BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 6),
-			BT_SDP_DATA_ELEM_LIST(
-			{
-				BT_SDP_TYPE_SIZE(BT_SDP_UUID16),
-				BT_SDP_ARRAY_16(BT_SDP_SERIAL_PORT_SVCLASS)
-			},
-			{
-				BT_SDP_TYPE_SIZE(BT_SDP_UINT16),
-				BT_SDP_ARRAY_16(0x0102)
-			},
-			)
-		},
-		)
-	),
-	BT_SDP_SERVICE_NAME("Serial Port"),
-};
-
-static struct bt_sdp_record spp_rec = BT_SDP_RECORD(spp_attrs);
-
-#endif /* CONFIG_BT_RFCOMM */
-
 static void cmd_auth_pincode(const struct shell *shell,
 			     size_t argc, char *argv[])
 {
@@ -357,140 +289,6 @@ static void cmd_l2cap_register(const struct shell *shell,
 		print(shell, "L2CAP psm %u registered\n", br_server.psm);
 	}
 }
-
-#if defined(CONFIG_BT_RFCOMM)
-static void rfcomm_recv(struct bt_rfcomm_dlc *dlci, struct net_buf *buf)
-{
-	print(NULL, "Incoming data dlc %p len %u\n", dlci, buf->len);
-}
-
-static void rfcomm_connected(struct bt_rfcomm_dlc *dlci)
-{
-	print(NULL, "Dlc %p connected\n", dlci);
-}
-
-static void rfcomm_disconnected(struct bt_rfcomm_dlc *dlci)
-{
-	print(NULL, "Dlc %p disconnected\n", dlci);
-}
-
-static struct bt_rfcomm_dlc_ops rfcomm_ops = {
-	.recv		= rfcomm_recv,
-	.connected	= rfcomm_connected,
-	.disconnected	= rfcomm_disconnected,
-};
-
-static struct bt_rfcomm_dlc rfcomm_dlc = {
-	.ops = &rfcomm_ops,
-	.mtu = 30,
-};
-
-static int rfcomm_accept(struct bt_conn *conn, struct bt_rfcomm_dlc **dlc)
-{
-	print(NULL, "Incoming RFCOMM conn %p\n", conn);
-
-	if (rfcomm_dlc.session) {
-		error(NULL, "No channels available");
-		return -ENOMEM;
-	}
-
-	*dlc = &rfcomm_dlc;
-
-	return 0;
-}
-
-struct bt_rfcomm_server rfcomm_server = {
-	.accept = &rfcomm_accept,
-};
-
-static void cmd_rfcomm_register(const struct shell *shell,
-				size_t argc, char *argv[])
-{
-	int ret;
-
-	if (rfcomm_server.channel) {
-		error(shell, "Already registered\n");
-		return;
-	}
-
-	rfcomm_server.channel = BT_RFCOMM_CHAN_SPP;
-
-	ret = bt_rfcomm_server_register(&rfcomm_server);
-	if (ret < 0) {
-		error(shell, "Unable to register channel %x\n", ret);
-		rfcomm_server.channel = 0;
-	} else {
-		print(shell, "RFCOMM channel %u registered\n",
-		      rfcomm_server.channel);
-		bt_sdp_register_service(&spp_rec);
-	}
-}
-
-static void cmd_rfcomm_connect(const struct shell *shell,
-			       size_t argc, char *argv[])
-{
-	u8_t channel;
-	int err;
-
-	if (!default_conn) {
-		error(shell, "Not connected\n");
-		return;
-	}
-
-	if (!shell_cmd_precheck(shell, argc == 2, NULL, 0)) {
-		return;
-	}
-
-	channel = strtoul(argv[1], NULL, 16);
-
-	err = bt_rfcomm_dlc_connect(default_conn, &rfcomm_dlc, channel);
-	if (err < 0) {
-		error(shell, "Unable to connect to channel %d (err %u)\n",
-		      channel, err);
-	} else {
-		print(shell, "RFCOMM connection pending\n");
-	}
-}
-
-static void cmd_rfcomm_send(const struct shell *shell,
-			    size_t argc, char *argv[])
-{
-	u8_t buf_data[DATA_BREDR_MTU] = { [0 ... (DATA_BREDR_MTU - 1)] =
-					    0xff };
-	int ret, len, count = 1;
-	struct net_buf *buf;
-
-	if (argc > 1) {
-		count = strtoul(argv[1], NULL, 10);
-	}
-
-	while (count--) {
-		buf = bt_rfcomm_create_pdu(&data_pool);
-		/* Should reserve one byte in tail for FCS */
-		len = min(rfcomm_dlc.mtu, net_buf_tailroom(buf) - 1);
-
-		net_buf_add_mem(buf, buf_data, len);
-		ret = bt_rfcomm_dlc_send(&rfcomm_dlc, buf);
-		if (ret < 0) {
-			print(shell, "Unable to send: %d\n", -ret);
-			net_buf_unref(buf);
-			break;
-		}
-	}
-}
-
-static void cmd_rfcomm_disconnect(const struct shell *shell,
-				  size_t argc, char *argv[])
-{
-	int err;
-
-	err = bt_rfcomm_dlc_disconnect(&rfcomm_dlc);
-	if (err) {
-		print(shell, "Unable to disconnect: %u\n", -err);
-	}
-}
-
-#endif /* CONFIG_BT_RFCOMM) */
 
 static void cmd_discoverable(const struct shell *shell,
 			     size_t argc, char *argv[])
@@ -753,12 +551,6 @@ SHELL_CREATE_STATIC_SUBCMD_SET(br_cmds) {
 	SHELL_CMD(oob, NULL, NULL, cmd_oob),
 	SHELL_CMD(pscan, NULL, "value: on, off", cmd_connectable),
 	SHELL_CMD(sdp-find, NULL, "<HFPAG>", cmd_sdp_find_record),
-#if defined(CONFIG_BT_RFCOMM)
-	SHELL_CMD(rfcomm-register, NULL, "<channel>", cmd_rfcomm_register),
-	SHELL_CMD(rfcomm-connect, NULL, "<channel>", cmd_rfcomm_connect),
-	SHELL_CMD(rfcomm-send, NULL, "<number of packets>", cmd_rfcomm_send),
-	SHELL_CMD(rfcomm-disconnect, NULL, HELP_NONE, cmd_rfcomm_disconnect),
-#endif /* CONFIG_BT_RFCOMM */
 	SHELL_SUBCMD_SET_END
 };
